+++ src/Smartstore.Web/Areas/Admin/Models/Settings/CustomerUserSettingsModel.cs
@@ -244,6 +244,9 @@ namespace Smartstore.Admin.Models
             [LocalizedDisplay("*Privacy.SameSiteMode")]
             public SameSiteMode SameSiteMode { get; set; } = SameSiteMode.Lax;
 
+            [LocalizedDisplay("*Privacy.VisitorCookieExpirationDays")]
+            public int VisitorCookieExpirationDays { get; set; } = 365;
+
             [LocalizedDisplay("*Privacy.StoreLastIpAddress")]
             public bool StoreLastIpAddress { get; set; }
 
@@ -324,6 +327,8 @@ namespace Smartstore.Admin.Models
         public CustomerUserSettingsValidator()
         {
             RuleFor(x => x.CustomerSettings.PasswordMinLength).GreaterThanOrEqualTo(4);
+
+            RuleFor(x => x.PrivacySettings.VisitorCookieExpirationDays).GreaterThan(0);
         }
     }
 }